Fixed displaying `hat` when zero (used to show "Loading..." instead)
The `.gt(0)` comparison didn't work correctly with `hat` as it is zero
by default. This made the UI display "Loading..." indefinitely.
Considering the fact that you set the default state of everything
to `-1` anyway, changing the `.gt(0)` condition to `.gte(0)` solves
the issue.

In order to keep consistency I have changed it not only for `hat`, but
for all fields displayed by the <SystemStatus /> component. It probably
won't matter for them anyway as they will never have zero values.